pyproject.toml
[project]
name = "enterprise_mcp_server"
version = "0.1.0"
description = ""
authors = [
    {name = "georgiedekker",email = "georgiedekker@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12, <3.14"
dependencies = [
    "fastmcp (>=2.3.0,<3.0.0)",
    "fastapi[all] (>=0.115.12,<0.116.0)",
    "restrictedpython (>=8.0,<9.0)",
    "importlib (>=1.0.4,<2.0.0)",
    "uvicorn[standard] (>=0.34.2,<0.35.0)",
    "pathlib (>=1.0.1,<2.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "sse-starlette (>=2.3.4,<3.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "jsonify (>=0.5,<0.6)",
    "httpx (>=0.28.1,<0.29.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "redis (>=6.0.0,<7.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "mcp (>=1.8.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "enterprise_mcp_server", from = "enterprise_mcp_server"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
